Move {embd,rnn}_descr_t to layer_descr.h
