+++ deeplima/libs/nn/birnn_seq_cls/birnn_seq_classifier.h
@@ -22,60 +22,15 @@
 
 #include <tuple>
 #include <vector>
-#include <sstream>
 
 #include "static_graph/static_graph.h"
 #include "deeplima/torch_wrp/torch_matrix.h"
+#include "nn/common/layer_descr.h"
 
 namespace deeplima
 {
 namespace nets
 {
-
-struct embd_descr_t
-{
-  std::string m_name;
-  size_t m_dim;
-  unsigned char m_type; // 0 - raw, 1 - embd
-
-  embd_descr_t(const std::string& name, size_t dim, unsigned char type=1)
-    : m_name(name), m_dim(dim), m_type(type) { }
-
-  explicit embd_descr_t(const std::string& str)
-  {
-    from_string(str);
-  }
-
-  std::string to_string() const
-  {
-    std::ostringstream s;
-    s << m_name << " " << (int)m_dim << " " << (int)m_type;
-    return s.str();
-  }
-
-  void from_string(const std::string& str)
-  {
-    std::istringstream s(str);
-    s >> m_name;
-    s >> m_dim;
-    int t = 0;
-    s >> t;
-    m_type = t;
-  }
-
-  bool operator==(const embd_descr_t& other)
-  {
-    return (other.m_name == m_name) && (other.m_dim == m_dim) && (other.m_type == m_type);
-  }
-};
-
-struct rnn_descr_t
-{
-  size_t m_dim;
-
-  explicit rnn_descr_t(size_t dim)
-    : m_dim(dim) { }
-};
 
 struct task_stat_t
 {
@@ -107,6 +62,14 @@ public:
                       const std::vector<uint32_t>& classes)
     : StaticGraphImpl(dicts,
                       generate_script(embd_descr, rnn_descr, output_names, classes)),
+      m_embd_descr(embd_descr)
+  {
+  }
+
+  BiRnnClassifierImpl(DictsHolder&& dicts,
+                      const std::vector<embd_descr_t>& embd_descr,
+                      const std::string& script)
+    : StaticGraphImpl(dicts, script),
       m_embd_descr(embd_descr)
   {
   }
